# contracts/openapi/kph.openapi.yaml
openapi: 3.1.0
info:
  title: Co.op Food KPH API
  version: 1.0.0-foundation
  description: |
    Contract-first boundary for the Store PWA and backend foundation.
    Business dates use ISO `yyyy-mm-dd` on the wire and `dd/mm/yyyy` in the UI.
    The business timezone is `Asia/Ho_Chi_Minh`. Store and actor authority always
    comes from the authenticated server-side session, never from mutation bodies.
servers:
  - url: http://localhost:8080
    description: Local backend
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Create a server-side session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated session and CSRF token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/session:
    get:
      tags: [Auth]
      operationId: getSession
      summary: Read current actor, roles and store memberships
      responses:
        '200':
          description: Current authenticated session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Invalidate the current session
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '204':
          description: Session invalidated.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/auth/password/change:
    post:
      tags: [Auth]
      operationId: changeOwnPassword
      summary: Change the authenticated user's password
      description: |
        Verifies the current password, applies the published password policy,
        writes a new algorithm-prefixed hash and invalidates every session issued
        before this change. Password values are never returned or audited.
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed; the current session is invalidated.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/CredentialValidation'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: KPH_SESSION
  parameters:
...
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Authentication required or credentials invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: CSRF validation failed or the actor lacks authority.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    CredentialValidation:
      description: Current password or new-password policy validation failed; code identifies the rule.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 256
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
          minLength: 1
          maxLength: 256
        newPassword:
          type: string
          format: password
          description: 15–64 Unicode code points; the server evaluates the full value without trimming or normalization.
          minLength: 1
          maxLength: 256
    SessionResponse:
      type: object
      additionalProperties: false
      required: [user, csrfToken]
      properties:
        user:
          $ref: '#/components/schemas/SessionUser'
        csrfToken:
          type: string
    SessionUser:
      type: object
      additionalProperties: false
      required: [id, username, displayName, globalRoles, stores]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        globalRoles:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/GlobalRole'
        stores:
          type: array
          items:
            $ref: '#/components/schemas/StoreContext'
    GlobalRole:
      type: string
      enum: [CATALOG_ADMIN, CHAIN_ADMIN]
    StoreRole:
      type: string
      enum: [EMPLOYEE, STORE_MANAGER]
    StoreContext:
      type: object
      additionalProperties: false
      required: [id, code, name, role]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          description: String identifier; leading zeroes are significant.
        name:
          type: string
        role:
          $ref: '#/components/schemas/StoreRole'
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          description: Stable machine-readable application error code.
...